@@ -1,17 +1,28 @@
-import React, { Fragment } from 'react';
-import { Header } from './components/Layout/Header';
-import { Meals } from './components/Meals/Meals';
+import { Fragment, useState } from 'react';
+
+import Header from './components/Layout/Header';
+import Meals from './components/Meals/Meals';
+import Cart from './components/Cart/Cart';
 
 function App() {
+  const [cartIsShown, setCartIsShown] = useState(false);
+
+  const showCartHandler = () => {
+    setCartIsShown(true);
+  };
+
+  const hideCartHandler = () => {
+    setCartIsShown(false);
+  };
+
   return (
-    <div>
-      <Fragment>
-        <Header />
-        <main>
-          <Meals />
-        </main>
-      </Fragment>
-    </div>
+    <Fragment>
+      {cartIsShown && <Cart onClose={hideCartHandler} />}
+      <Header onShowCart={showCartHandler} />
+      <main>
+        <Meals />
+      </main>
+    </Fragment>
   );
 }
 
