@@ -1,10 +1,9 @@
-import React from 'react';
-import { CartIcon } from '../Cart/CartIcon';
+import CartIcon from '../Cart/CartIcon';
 import classes from './HeaderCartButton.module.css';
 
-export const HeaderCartButton = () => {
+const HeaderCartButton = props => {
   return (
-    <button className={classes.button}>
+    <button className={classes.button} onClick={props.onClick}>
       <span className={classes.icon}>
         <CartIcon />
       </span>
@@ -13,3 +12,5 @@ export const HeaderCartButton = () => {
     </button>
   );
 };
+
+export default HeaderCartButton;
