@@ -1,9 +1,9 @@
-import React from 'react';
+import MealItemForm from './MealItemForm';
 import classes from './MealItem.module.css';
-import { MealItemForm } from './MealItemForm';
 
-export const MealItem = props => {
+const MealItem = props => {
   const price = `$${props.price.toFixed(2)}`;
+
   return (
     <li className={classes.meal}>
       <div>
@@ -12,8 +12,10 @@ export const MealItem = props => {
         <div className={classes.price}>{price}</div>
       </div>
       <div>
-        <MealItemForm />
+        <MealItemForm id={props.id} />
       </div>
     </li>
   );
 };
+
+export default MealItem;
