@@ -1,14 +1,13 @@
-import React from 'react';
-import { Input } from '../../UI/Input';
+import Input from '../../UI/Input';
 import classes from './MealItemForm.module.css';
 
-export const MealItemForm = props => {
+const MealItemForm = props => {
   return (
     <form className={classes.form}>
       <Input
         label='Amount'
         input={{
-          id: 'amount',
+          id: 'amount_' + props.id,
           type: 'number',
           min: '1',
           max: '5',
@@ -20,3 +19,5 @@ export const MealItemForm = props => {
     </form>
   );
 };
+
+export default MealItemForm;
