@@ -1,8 +1,9 @@
-import React, { Fragment } from 'react';
-import { MealsSummary } from './MealsSummary';
-import { AvailableMeals } from './AvailableMeals';
+import { Fragment } from 'react';
 
-export const Meals = () => {
+import MealsSummary from './MealsSummary';
+import AvailableMeals from './AvailableMeals';
+
+const Meals = () => {
   return (
     <Fragment>
       <MealsSummary />
@@ -10,3 +11,5 @@ export const Meals = () => {
     </Fragment>
   );
 };
+
+export default Meals;
