@@ -1,10 +1,9 @@
-import React from 'react';
 import classes from './MealsSummary.module.css';
 
-export const MealsSummary = () => {
+const MealsSummary = () => {
   return (
     <section className={classes.summary}>
-      <h2>Delicious Food, Delivered to You</h2>
+      <h2>Delicious Food, Delivered To You</h2>
       <p>
         Choose your favorite meal from our broad selection of available meals
         and enjoy a delicious lunch or dinner at home.
@@ -16,3 +15,5 @@ export const MealsSummary = () => {
     </section>
   );
 };
+
+export default MealsSummary;
