@@ -1,7 +1,6 @@
-import React from 'react';
 import classes from './Input.module.css';
 
-export const Input = props => {
+const Input = (props) => {
   return (
     <div className={classes.input}>
       <label htmlFor={props.input.id}>{props.label}</label>
@@ -9,3 +8,5 @@ export const Input = props => {
     </div>
   );
 };
+
+export default Input;
